mixture data from results and updated version
The alternative mixture data has been removed from the API result examples and the type definition in the code. This was an unnecessary set of data that was not required and thereby has been eliminated to keep the code cleaner and more efficient. Also, the version has been updated to 0.30 from 0.29 as this constitutes a significant change in the data structure.