Refactor datasheet handling and introduce MixerMeta type
This commit refactors the way datasheets are handled by introducing a new MixerMeta type. Datasheets are now integrated into this new MixerMeta type which is used in multiple areas, including the API type definitions and example payloads. Furthermore, various new properties and types related to rule conditions, HubSpot forms, and more are introduced for better structure and flexibility. The datasheet content is enriched with rules that dictate its visibility based on set properties.